forge fmt

+++ src/ICreateProject.sol
@@ -7,40 +7,29 @@ interface ICreateProject {
     function create(
         string memory _projectName,
         string memory _projectSymbol,
-        uint _maxTokenSupply,
-        uint _minTokenSale,
-        uint _price,
-        uint _publicSale,
-        uint _amountSteps,
-        uint[] memory _timeSteps
+        uint256 _maxTokenSupply,
+        uint256 _minTokenSale,
+        uint256 _price,
+        uint256 _publicSale,
+        uint256 _amountSteps,
+        uint256[] memory _timeSteps
     ) external returns (uint32, uint32, uint32);
 
     function ownerOfProject(uint32 projectId) external view returns (address);
 
     function isPublicSale(uint32 projectId) external view returns (bool);
 
-    function projectsViewSteps(
-        uint32 projectId
-    )
+    function projectsViewSteps(uint32 projectId)
         external
         view
-        returns (
-            uint8,
-            uint[] memory,
-            uint[] memory,
-            uint[] memory,
-            bool[] memory,
-            uint,
-            bool
-        );
+        returns (uint8, uint256[] memory, uint256[] memory, uint256[] memory, bool[] memory, uint256, bool);
 
-    function projectsViewPrice(
-        uint32 projectId
-    ) external view returns (uint, uint, uint, uint, uint, bool, bool, uint);
+    function projectsViewPrice(uint32 projectId)
+        external
+        view
+        returns (uint256, uint256, uint256, uint256, uint256, bool, bool, uint256);
 
-    function projectsViewMain(
-        uint32 projectId
-    )
+    function projectsViewMain(uint32 projectId)
         external
         view
         returns (ProjectToken, address, string memory, string memory);
@@ -48,44 +37,32 @@ interface ICreateProject {
     function updateSteps(
         uint32 _projectIdSteps,
         uint8 _amountSteps,
-        uint[] memory _timeSteps,
-        uint[] memory _rewardTimePerStep,
-        uint[] memory _dateSteps,
+        uint256[] memory _timeSteps,
+        uint256[] memory _rewardTimePerStep,
+        uint256[] memory _dateSteps,
         bool[] memory _isStepsPerProject,
-        uint _startTime,
+        uint256 _startTime,
         bool _isPublicSale
     ) external;
 
-    function orderSteps(
-        uint32 _projectIdSteps,
-        bool[] memory _isStepsPerProject
-    ) external;
+    function orderSteps(uint32 _projectIdSteps, bool[] memory _isStepsPerProject) external;
 
     function updatePrice(
         uint32 _projectIdPrice,
-        uint _tokenSupply,
-        uint _maxTokenSupply,
-        uint _minTokenSale,
-        uint _price,
-        uint _publicSale,
+        uint256 _tokenSupply,
+        uint256 _maxTokenSupply,
+        uint256 _minTokenSale,
+        uint256 _price,
+        uint256 _publicSale,
         bool _isProjectAlive,
-        uint _fundsForProject
+        uint256 _fundsForProject
     ) external;
 
-    function orderPrice(
-        uint32 _projectIdPrice,
-        uint _tokenSupply,
-        uint _fundsForProject
-    ) external;
+    function orderPrice(uint32 _projectIdPrice, uint256 _tokenSupply, uint256 _fundsForProject) external;
 
-    function witchStepAlive(
-        uint32 _projectId
-    ) external view returns (uint8 step);
+    function witchStepAlive(uint32 _projectId) external view returns (uint8 step);
 
-    function updateAfterSBSFund(
-        uint32 _projectIdSteps,
-        uint8 _stepIsLive
-    ) external;
+    function updateAfterSBSFund(uint32 _projectIdSteps, uint8 _stepIsLive) external;
 
     function closeProject(uint32 _projectIdPrice) external;
