+++ src/ProjectManagerPrice.sol
@@ -7,33 +7,35 @@ contract ProjectManagerPrice {
     address public tokensMainForProjectAddr;
 
     struct CreatePrice {
-        uint tokenSupply; // минтится в процессе
-        uint maxTokenSupply; // устанавливается при создании, включает все сегменты
-        uint minTokenSale; // устанавливается при создании, сколько нужно продать первично
-        uint price;
-        uint publicSale; // сколько хотят продать монет
+        uint256 tokenSupply; // минтится в процессе
+        uint256 maxTokenSupply; // устанавливается при создании, включает все сегменты
+        uint256 minTokenSale; // устанавливается при создании, сколько нужно продать первично
+        uint256 price;
+        uint256 publicSale; // сколько хотят продать монет
         bool isProjectAlive; // прошел ли проект голосование либо полностью завершен
         bool isPrijectGetAllTokens; // получил ли проект все свои токены после завершения работы
-        uint fundsForProject; // общее кол-во средств для проекта
+        uint256 fundsForProject; // общее кол-во средств для проекта
     }
+
     event ProjectCreatedPrice(CreatePrice project);
+
     uint32 public projectIdPrice = 1;
 
     mapping(uint32 => CreatePrice) getProjectPrice;
 
-    function setUsingAddrsForPrice(
-        address _orderingAddr,
-        address _votingAddr,
-        address _tokensMainForProjectAddr
-    ) internal {
+    function setUsingAddrsForPrice(address _orderingAddr, address _votingAddr, address _tokensMainForProjectAddr)
+        internal
+    {
         orderingAddr = _orderingAddr;
         votingAddr = _votingAddr;
         tokensMainForProjectAddr = _tokensMainForProjectAddr;
     }
 
-    function projectsViewPrice(
-        uint32 projectId
-    ) public view returns (uint, uint, uint, uint, uint, bool, bool, uint) {
+    function projectsViewPrice(uint32 projectId)
+        public
+        view
+        returns (uint256, uint256, uint256, uint256, uint256, bool, bool, uint256)
+    {
         CreatePrice memory project = getProjectPrice[projectId];
         return (
             project.tokenSupply,
@@ -47,12 +49,10 @@ contract ProjectManagerPrice {
         );
     }
 
-    function createProjectPrice(
-        uint _maxTokenSupply,
-        uint _minTokenSale,
-        uint _price,
-        uint _publicSale
-    ) internal returns (uint32) {
+    function createProjectPrice(uint256 _maxTokenSupply, uint256 _minTokenSale, uint256 _price, uint256 _publicSale)
+        internal
+        returns (uint32)
+    {
         CreatePrice memory project = CreatePrice({
             tokenSupply: 0,
             maxTokenSupply: _maxTokenSupply,
@@ -70,11 +70,7 @@ contract ProjectManagerPrice {
         return projectIdPrice - 1;
     }
 
-    function orderPrice(
-        uint32 _projectIdPrice,
-        uint _tokenSupply,
-        uint _fundsForProject
-    ) external {
+    function orderPrice(uint32 _projectIdPrice, uint256 _tokenSupply, uint256 _fundsForProject) external {
         require(msg.sender == orderingAddr, "Not allowed");
         getProjectPrice[_projectIdPrice].tokenSupply = _tokenSupply;
         getProjectPrice[_projectIdPrice].fundsForProject = _fundsForProject;
