+++ src/Voting.sol
@@ -16,33 +16,17 @@ contract Voting {
     // голосование за каждый этап, начиная со 2-го, т.к. 1-й прошел после удачного сбора средств
     uint32 votingTime = 604800; // 7 days
     uint8 negativeShare = 51; // 51% против, чтобы проект перестал получать финансирование
-    mapping(uint32 => mapping(uint8 => uint)) public projectStepsResultVoting; // результаты прохождения этапов, сколько голосов против, начиная с 2: (т.к. 0: паблик, 1: 1-я работа)
-    mapping(uint32 => mapping(uint8 => mapping(address => bool)))
-        public isVoting; // голосовал уже или нет
+    mapping(uint32 => mapping(uint8 => uint256)) public projectStepsResultVoting; // результаты прохождения этапов, сколько голосов против, начиная с 2: (т.к. 0: паблик, 1: 1-я работа)
+    mapping(uint32 => mapping(uint8 => mapping(address => bool))) public isVoting; // голосовал уже или нет
 
-    function viewProjectResultVoting(
-        uint32 _projectId,
-        uint8 _step
-    ) public view returns (uint) {
+    function viewProjectResultVoting(uint32 _projectId, uint8 _step) public view returns (uint256) {
         return projectStepsResultVoting[_projectId][_step];
     }
 
     function vote(uint32 _projectId) public {
-        uint userOrderedTokens = ordering_contract.getUserOrderedTokens(
-            _projectId,
-            msg.sender
-        );
+        uint256 userOrderedTokens = ordering_contract.getUserOrderedTokens(_projectId, msg.sender);
         require(userOrderedTokens != 0, "You haven't the tokens to vote!");
-        (
-            uint tokenSupply,
-            ,
-            ,
-            ,
-            ,
-            bool isProjectAlive,
-            ,
-
-        ) = createProject_contract.projectsViewPrice(_projectId);
+        (uint256 tokenSupply,,,,, bool isProjectAlive,,) = createProject_contract.projectsViewPrice(_projectId);
         require(isProjectAlive, "Project is over!");
 
         uint8 stepIsLive = createProject_contract.witchStepAlive(_projectId);
@@ -59,37 +43,17 @@ contract Voting {
         // stepIsLive[4] = false;  // stepIsLive[4] = false;  // stepIsLive[4] = false;  // stepIsLive[4] = false;   // stepIsLive[4] = true;
         require(stepIsLive > 1, "Voting hasn't started yet!"); // проверить общую длину массива stepIsLive
 
-        (
-            uint amountSteps,
-            ,
-            ,
-            uint[] memory dateSteps,
-            ,
-            ,
-
-        ) = createProject_contract.projectsViewSteps(_projectId);
+        (uint256 amountSteps,,, uint256[] memory dateSteps,,,) = createProject_contract.projectsViewSteps(_projectId);
         require(stepIsLive < amountSteps, "All votes is done!");
-        require(
-            block.timestamp > dateSteps[stepIsLive],
-            "Voting hasn't started yet!"
-        );
-        require(
-            block.timestamp <= dateSteps[stepIsLive] + votingTime,
-            "Voting is over or not start yet!"
-        );
-        require(
-            !isVoting[_projectId][stepIsLive][msg.sender],
-            "You've already voted!"
-        );
+        require(block.timestamp > dateSteps[stepIsLive], "Voting hasn't started yet!");
+        require(block.timestamp <= dateSteps[stepIsLive] + votingTime, "Voting is over or not start yet!");
+        require(!isVoting[_projectId][stepIsLive][msg.sender], "You've already voted!");
         isVoting[_projectId][stepIsLive][msg.sender] = true;
 
-        uint userVotingPower = userOrderedTokens;
+        uint256 userVotingPower = userOrderedTokens;
         projectStepsResultVoting[_projectId][stepIsLive] += userVotingPower;
 
-        if (
-            projectStepsResultVoting[_projectId][stepIsLive] >
-            (tokenSupply / 100) * negativeShare
-        ) {
+        if (projectStepsResultVoting[_projectId][stepIsLive] > (tokenSupply / 100) * negativeShare) {
             createProject_contract.closeProject(_projectId);
         }
     }
